+++ resolwe/flow/serializers.py
@@ -150,13 +150,15 @@ class DataSerializer(ResolweBaseSerializer):
 
         model = Data
         update_protected_fields = ('contributor', 'process',)
-        read_only_fields = ('id', 'created', 'modified', 'started', 'finished', 'checksum',
-                            'status', 'process_progress', 'process_rc', 'process_info',
-                            'process_warning', 'process_error', 'process_type',
-                            'process_input_schema', 'process_output_schema',
-                            'process_slug', 'process_name', 'descriptor_dirty')
-        fields = ('slug', 'name', 'contributor', 'input', 'output', 'descriptor_schema',
-                  'descriptor', 'tags') + update_protected_fields + read_only_fields
+        read_only_fields = (
+            'id', 'created', 'modified', 'started', 'finished', 'checksum', 'status', 'process_progress', 'process_rc',
+            'process_info', 'process_warning', 'process_error', 'process_type', 'process_input_schema',
+            'process_output_schema', 'process_slug', 'process_name', 'descriptor_dirty',
+        )
+        fields = (
+            'slug', 'name', 'contributor', 'input', 'output', 'descriptor_schema', 'descriptor', 'tags',
+            'process_memory', 'process_cores',
+        ) + update_protected_fields + read_only_fields
 
     def get_fields(self):
         """Dynamically adapt fields based on the current request."""